+++ src/libai/nn/supervised/RBF.java
@@ -193,7 +193,7 @@ public class RBF extends Adaline {
 
 					exit = false;
 				} else {
-					aux1.multiply(1.0 / (double) total, aux);
+					aux1.multiply(1.0 / total, aux);
 
 					if (!(aux.equals(ctemp[i]))) {
 						aux.copy(ctemp[i]);
